refactor settings to be custom alert dialog
