+++ app/src/main/java/com/example/csci4176_groupproject/Level4Activity.kt
@@ -60,40 +60,7 @@ class Level4Activity : AppCompatActivity() {
 
         val settingsButton = findViewById<ImageButton>(R.id.SettingsButton)
         settingsButton.setOnClickListener {
-
-            val builder = AlertDialog.Builder(this, R.style.SettingsDialog)
-                .create()
-            val view = layoutInflater.inflate(R.layout.settings_dialog, null)
-            builder.setView(view)
-
-            // set the state of the settings
-            val colorBlindModeView = view.findViewById<Switch>(R.id.colorBlindSwitch)
-            colorBlindModeView.isChecked = settingPrefs.getBoolean("colorBlind", false)
-            val hapticsSwitchView = view.findViewById<Switch>(R.id.hapticsSwitch)
-            hapticsSwitchView.isChecked = settingPrefs.getBoolean("haptics", true)
-            val soundView = view.findViewById<ToggleButton>(R.id.soundToggle)
-            soundView.isChecked = settingPrefs.getBoolean("sound", false)
-
-            val applyButton = view.findViewById<Button>(R.id.applyButton)
-            applyButton.setOnClickListener {
-
-                // save state of settings
-                val editor: SharedPreferences.Editor = settingPrefs.edit()
-                editor.putBoolean("colorBlind", colorBlindModeView.isChecked)
-                editor.putBoolean("haptics", hapticsSwitchView.isChecked)
-                editor.putBoolean("sound", soundView.isChecked)
-                editor.apply()
-
-                builder.dismiss()
-            }
-            val cancelButton = view.findViewById<Button>(R.id.cancelButton)
-            cancelButton.setOnClickListener {
-//              if the cancel button is hit, don't save settings and exit
-                builder.cancel()
-            }
-//          the user must hit either the cancel or apply button to close the dialog
-            builder.setCanceledOnTouchOutside(false)
-            builder.show()
+            settingsDialog(context = this).showSettings()
         }
 
         // Setup wall tiles
